agrego exportacion de constante y cambio tipo de dato del input de whatsapp
