+++ catering.html
@@ -1,122 +1,135 @@
 <!DOCTYPE html>
 <html lang="es">
+  <head>
+    <meta charset="UTF-8" />
+    <meta name="viewport" content="width=device-width, initial-scale=1.0" />
+    <title>IDW S.A. - Salones de fiestas infantiles</title>
+    <link rel="icon" href="imagenes/Logo-idw.png" type="image/png" />
 
-<head>
-  <meta charset="UTF-8" />
-  <meta name="viewport" content="width=device-width, initial-scale=1.0" />
-  <title>IDW S.A. - Salones de fiestas infantiles</title>
-  <link rel="icon" href="imagenes/Logo-idw.png" type="image/png" />
+    <link rel="stylesheet" href="src/styles/estilos.css" />
+    <link
+      rel="stylesheet"
+      href="src/bootstrap-5.0.2/dist/css/bootstrap.min.css"
+    />
+  </head>
 
-  <link rel="stylesheet" href="src/styles/estilos.css" />
-  <link rel="stylesheet" href="src/bootstrap-5.0.2/dist/css/bootstrap.min.css" />
+  <body class="bg-body">
+    <!--Header-->
+    <div id="header-placeholder"></div>
 
-</head>
+    <main class="d-block justify-content-center px-3 m-1">
+      <!--Bienvenida-->
+      <section
+        id="conocenos"
+        class="bg-danger text-center shadow mt-4 py-5 px-3 rounded-3"
+      >
+        <div class="container">
+          <h2 class="display-2 fw-bold">Catering</h2>
+          <p class="mx-auto fs-5">
+            La planificación de un catering para eventos suele incluir la
+            participación de un coordinador encargado de supervisar toda la
+            logística, así como también mozos, equipo de cocina y personal
+            responsable del montaje del servicio.
+          </p>
+        </div>
+      </section>
 
-<body class="bg-body">
-  <!--Header-->
-  <div id="header-placeholder"></div>
+      <hr class="bg-dark my-4 mt-5" />
 
-  <main class="d-block justify-content-center px-3 m-1">
-    <!--Bienvenida-->
-    <section id="conocenos" class="bg-danger text-center shadow mt-4 py-5 px-3 rounded-3">
-      <div class="container">
-        <h2 class="display-2 fw-bold">Catering</h2>
-        <p class="mx-auto fs-5">
-          La planificación de un catering para eventos suele incluir la participación de un coordinador encargado de
-          supervisar toda la logística, así como también mozos, equipo de cocina y personal responsable del montaje
-          del
-          servicio.
-        </p>
-      </div>
-    </section>
-
-    <hr class="bg-dark my-4 mt-5" />
-
-
-    <!-- Menús -->
-    <section id="catalogo-salones" class="text-info">
-
-      <!-- Salon 1 -->
-      <div class="salones">
-        <h3>Piccolino Eventos</h3>
-        <div class="salon">
-          <div>
-            <img src="src/imagenes/Catering-Piccolino-Eventos.jpg" alt="Piedra Libre" />
-          </div>
-          <div>
-            <p>
-              <strong>Dirección:</strong>
-              <a href="https://g.co/kgs/TmDPECa" target="_blank">
-                Av. Siempre Viva 123, Springfield - Oregón.
-              </a>
-            </p>
-            <p>
-              <strong>Contacto:</strong>
-              <a href="https://wa.me/543435194040" target="_blank">Whatsapp</a>
-            </p>
+      <!-- Menús -->
+      <section id="catalogo-salones" class="text-info">
+        <!-- Salon 1 -->
+        <div class="salones">
+          <h3>Piccolino Eventos</h3>
+          <div class="salon">
+            <div>
+              <img
+                src="src/imagenes/Catering-Piccolino-Eventos.jpg"
+                alt="Piedra Libre"
+              />
+            </div>
+            <div>
+              <p>
+                <strong>Dirección:</strong>
+                <a href="https://g.co/kgs/TmDPECa" target="_blank">
+                  Av. Siempre Viva 123, Springfield - Oregón.
+                </a>
+              </p>
+              <p>
+                <strong>Contacto:</strong>
+                <a href="https://wa.me/543435194040" target="_blank"
+                  >Whatsapp</a
+                >
+              </p>
+            </div>
           </div>
         </div>
-      </div>
 
-      <hr class="bg-dark mx-5 my-4" />
+        <hr class="bg-dark mx-5 my-4" />
 
-
-      <!-- Salon 2 -->
-      <div class="salones">
-        <h3>Tutto Speciale Catering</h3>
-        <div class="salon">
-          <div>
-            <img src="src/imagenes/tutto-speciale-catering.png" alt="Cachetes Felices" />
-          </div>
-          <div>
-            <p>
-              <strong>Dirección:</strong>
-              <a href="https://g.co/kgs/dPGFGtu" target="_blank">
-                Av. Francisco Ramirez 5080, Paraná - Entre Ríos.
-              </a>
-            </p>
-            <p>
-              <strong>Contacto:</strong>
-              <a href="https://wa.me/543435728358" target="_blank">Whatsapp</a>
-            </p>
-          </div>
-        </div>
-      </div>
-
-      <hr class="bg-dark mx-5 my-4" />
-
-      <!-- Salon 3 -->
-      <div class="salones">
-        <h3>Aire Catering</h3>
-        <div class="salon">
-          <div>
-            <img src="src/imagenes/Catering-para-eventos-sociales-Aire-Catering.jpg" alt="Garden Park" />
-          </div>
-          <div>
-            <p>
-              <strong>Dirección:</strong>
-              <a href="https://g.co/kgs/zwKVW1J" target="_blank">
-                Calle Falsa 123, Springfield - Oregón.
-              </a>
-            </p>
-            <p>
-              <strong>Contacto:</strong>
-              <a href="https://wa.me/543434402888" target="_blank">Whatsapp</a>
-            </p>
+        <!-- Salon 2 -->
+        <div class="salones">
+          <h3>Tutto Speciale Catering</h3>
+          <div class="salon">
+            <div>
+              <img
+                src="src/imagenes/tutto-speciale-catering.png"
+                alt="Cachetes Felices"
+              />
+            </div>
+            <div>
+              <p>
+                <strong>Dirección:</strong>
+                <a href="https://g.co/kgs/dPGFGtu" target="_blank">
+                  Av. Francisco Ramirez 5080, Paraná - Entre Ríos.
+                </a>
+              </p>
+              <p>
+                <strong>Contacto:</strong>
+                <a href="https://wa.me/543435728358" target="_blank"
+                  >Whatsapp</a
+                >
+              </p>
+            </div>
           </div>
         </div>
-      </div>
 
-    </section>
+        <hr class="bg-dark mx-5 my-4" />
 
-  </main>
+        <!-- Salon 3 -->
+        <div class="salones">
+          <h3>Aire Catering</h3>
+          <div class="salon">
+            <div>
+              <img
+                src="src/imagenes/Catering-para-eventos-sociales-Aire-Catering.jpg"
+                alt="Garden Park"
+              />
+            </div>
+            <div>
+              <p>
+                <strong>Dirección:</strong>
+                <a href="https://g.co/kgs/zwKVW1J" target="_blank">
+                  Calle Falsa 123, Springfield - Oregón.
+                </a>
+              </p>
+              <p>
+                <strong>Contacto:</strong>
+                <a href="https://wa.me/543434402888" target="_blank"
+                  >Whatsapp</a
+                >
+              </p>
+            </div>
+          </div>
+        </div>
+      </section>
+    </main>
 
-  <!--Footer-->
-  <div id="footer"></div>
+    <!--Footer-->
+    <div id="footer"></div>
 
-  <script src="src/scripts/main.js" defer></script>
-  <script src="src/bootstrap-5.0.2/dist/js/bootstrap.bundle.min.js" defer></script>
-  <script src="src/bootstrap-5.0.2/dist/js/bootstrap.min.js" defer></script>
-</body>
-
+    <script src="src/scripts/main.js" type="module"></script>
+    <script src="src/bootstrap-5.0.2/dist/js/bootstrap.bundle.min.js"></script>
+    <script src="src/bootstrap-5.0.2/dist/js/bootstrap.min.js"></script>
+  </body>
 </html>
