+++ contacto.html
@@ -1,51 +1,54 @@
 <!DOCTYPE html>
 <html lang="es">
+  <head>
+    <meta charset="UTF-8" />
+    <meta name="viewport" content="width=device-width, initial-scale=1.0" />
+    <title>IDW S.A. - Salones de fiestas infantiles</title>
+    <link rel="icon" href="imagenes/Logo-idw.png" type="image/png" />
 
-<head>
-  <meta charset="UTF-8" />
-  <meta name="viewport" content="width=device-width, initial-scale=1.0" />
-  <title>IDW S.A. - Salones de fiestas infantiles</title>
-  <link rel="icon" href="imagenes/Logo-idw.png" type="image/png" />
+    <link rel="stylesheet" href="src/styles/estilos.css" />
+    <link
+      rel="stylesheet"
+      href="src/bootstrap-5.0.2/dist/css/bootstrap.min.css"
+    />
+  </head>
 
-  <link rel="stylesheet" href="src/styles/estilos.css" />
-  <link rel="stylesheet" href="src/bootstrap-5.0.2/dist/css/bootstrap.min.css" />
+  <body>
+    <!--Header-->
+    <div id="header-placeholder"></div>
 
-</head>
-
-<body>
-  <!--Header-->
-  <div id="header-placeholder"></div>
-
-  <main class="container min-vh-100 d-flex justify-content-center align-items-center">
-  <div class="container-sm" style="max-width: 720px;">
-  <!-- <div class="container-sm"> -->
-    <div class="m-4 p-4 shadow rounded-3">
-      <h2 class="text-center text-info mb-4">Contacto</h2>
-      <form>
-        <div class="mb-3">
-          <label for="nombre" class="form-label">Nombre</label>
-          <input type="text" class="form-control" id="nombre" required>
+    <main
+      class="container min-vh-100 d-flex justify-content-center align-items-center"
+    >
+      <div class="container-sm" style="max-width: 720px">
+        <!-- <div class="container-sm"> -->
+        <div class="m-4 p-4 shadow rounded-3">
+          <h2 class="text-center text-info mb-4">Contacto</h2>
+          <form>
+            <div class="mb-3">
+              <label for="nombre" class="form-label">Nombre</label>
+              <input type="text" class="form-control" id="nombre" required />
+            </div>
+            <div class="mb-3">
+              <label for="email" class="form-label">Email</label>
+              <input type="email" class="form-control" id="email" required />
+            </div>
+            <div class="mb-3">
+              <label for="telefono" class="form-label">Teléfono</label>
+              <input type="tel" class="form-control" id="telefono" />
+            </div>
+            <div class="mb-3">
+              <label for="mensaje" class="form-label">Mensaje</label>
+              <textarea class="form-control" id="mensaje" rows="3"></textarea>
+            </div>
+            <div class="text-center">
+              <button type="submit" class="btn btn-success">Enviar</button>
+            </div>
+          </form>
         </div>
-        <div class="mb-3">
-          <label for="email" class="form-label">Email</label>
-          <input type="email" class="form-control" id="email" required>
-        </div>
-        <div class="mb-3">
-          <label for="telefono" class="form-label">Teléfono</label>
-          <input type="tel" class="form-control" id="telefono">
-        </div>
-        <div class="mb-3">
-          <label for="mensaje" class="form-label">Mensaje</label>
-          <textarea class="form-control" id="mensaje" rows="3"></textarea>
-        </div>
-        <div class="text-center">
-          <button type="submit" class="btn btn-success">Enviar</button>
-        </div>
-      </form>
-    </div>
-  </div>
+      </div>
 
-    <!-- 
+      <!-- 
     <section id="contacto">
       <div class="contacto">
         <h2>Contacto</h2>
@@ -75,16 +78,12 @@
         </form>
       </div>
     </section> -->
+    </main>
 
+    <!--Footer-->
+    <div id="footer"></div>
 
-  </main>
-
-  <!--Footer-->
-  <div id="footer"></div>
-
-
-  <script src="src/scripts/main.js" defer></script>
-  <script src="src/bootstrap-5.0.2/dist/js/bootstrap.min.js" defer></script>
-</body>
-
+    <script src="src/scripts/main.js" type="module"></script>
+    <script src="src/bootstrap-5.0.2/dist/js/bootstrap.min.js"></script>
+  </body>
 </html>
