+++ index.html
@@ -193,7 +193,7 @@
     <!--Footer-->
     <div id="footer"></div>
 
-    <script src="src/scripts/main.js" defer></script>
+    <script src="src/scripts/main.js" type="module"></script>
     <script src="src/bootstrap-5.0.2/dist/js/bootstrap.bundle.min.js" defer></script>
     <script src="src/bootstrap-5.0.2/dist/js/bootstrap.min.js" defer></script>
