+++ src/components/header.html
@@ -1,4 +1,4 @@
-<header class="navbar navbar-expand-lg navbar-light bg-light bg-gradient" >
+<header class="navbar navbar-expand-lg navbar-light bg-light bg-gradient">
   <div class="container-fluid">
     <!-- LOGO -->
     <a class="navbar-brand" href="/">
@@ -6,20 +6,42 @@
     </a>
 
     <!-- BOTÓN HAMBURGUESA PARA MÓVIL -->
-    <button class="navbar-toggler shadow" type="button" data-bs-toggle="collapse" data-bs-target="#navbarMenu"
-      aria-controls="navbarMenu" aria-expanded="false" aria-label="Toggle navigation">
+    <button
+      class="navbar-toggler shadow"
+      type="button"
+      data-bs-toggle="collapse"
+      data-bs-target="#navbarMenu"
+      aria-controls="navbarMenu"
+      aria-expanded="false"
+      aria-label="Toggle navigation"
+    >
       <span class="navbar-toggler-icon"></span>
     </button>
 
     <!-- MENÚ DE NAVEGACIÓN -->
-    <nav class="collapse navbar-collapse justify-content-end mr-4" id="navbarMenu">
+    <nav
+      class="collapse navbar-collapse justify-content-end mr-4"
+      id="navbarMenu"
+    >
       <ul class="navbar-nav mb-2 mb-lg-0">
-        <li class="nav-item"><a class="nav-link fw-bold" href="/">Inicio</a></li>
-        <li class="nav-item"><a class="nav-link fw-bold" href="eventos.html">Salones de Eventos</a></li>
-        <li class="nav-item"><a class="nav-link fw-bold" href="catering.html">Catering</a></li>
-        <li class="nav-item"><a class="nav-link fw-bold" href="conocenos.html">Conócenos</a></li>
-        <li class="nav-item"><a class="nav-link fw-bold" href="contacto.html">Contacto</a></li>
-        <!-- <li class="nav-item "><a class="nav-link" href="dashBoard.html">Dashboard</a></li> -->
+        <li class="nav-item">
+          <a class="nav-link fw-bold" href="/">Inicio</a>
+        </li>
+        <li class="nav-item">
+          <a class="nav-link fw-bold" href="eventos.html">Salones de Eventos</a>
+        </li>
+        <li class="nav-item">
+          <a class="nav-link fw-bold" href="catering.html">Catering</a>
+        </li>
+        <li class="nav-item">
+          <a class="nav-link fw-bold" href="conocenos.html">Conócenos</a>
+        </li>
+        <li class="nav-item">
+          <a class="nav-link fw-bold" href="contacto.html">Contacto</a>
+        </li>
+        <li class="nav-item">
+          <a class="nav-link" href="dashBoard.html">Dashboard</a>
+        </li>
       </ul>
     </nav>
   </div>
