--- src/data/db.js
@@ -1,52 +1,0 @@
-const salonesIniciales = [
-  {
-    id: 1,
-    titulo: "Piedra Libre Eventos",
-    descripcion: "Salón para fiestas infantiles",
-    direccion: "Pascual Palma 857, Paraná - Entre Ríos",
-    whatsapp: "543435194040",
-    imagen: "src/imagenes/piedra-libre-1.jpg"
-  },
-  {
-    id: 2,
-    titulo: "Cachetes Felices",
-    descripcion: "Salón para fiestas infantiles",
-    direccion: "Av. Francisco Ramirez 5080, Paraná - Entre Ríos",
-    whatsapp: "543435728358",
-    imagen: "src/imagenes/Cachetes-felices-1.jpg"
-  },
-  {
-    id: 3,
-    titulo: "Garden Park",
-    descripcion: "Salón para fiestas infantiles",
-    direccion: "Courreges 345, Paraná - Entre Ríos",
-    whatsapp: "543434402888",
-    imagen: "src/imagenes/garden-park-1.jpg"
-  },
-  {
-    id: 4,
-    titulo: "Safari Eventos",
-    descripcion: "Salón para fiestas infantiles",
-    direccion: "Cervantes 273, Paraná - Entre Ríos",
-    whatsapp: "543435161859",
-    imagen: "src/imagenes/safari-1.jpg"
-  },
-  {
-    id: 5,
-    titulo: "Ciudad Mágica",
-    descripcion: "Salón para fiestas infantiles",
-    direccion: "Rondeau 971, Paraná - Entre Ríos",
-    whatsapp: "543435224541",
-    imagen: "src/imagenes/ciudad-magica-1.jpg"
-  },
-  {
-    id: 6,
-    titulo: "Bob Esponja",
-    descripcion: "Salón para fiestas infantiles",
-    direccion: "Gral. Gervasio Artigas 1225, Paraná - Entre Ríos",
-    whatsapp: "543434531818",
-    imagen: "src/imagenes/bob-esponja-1.jpg"
-  }
-];
-
-export { salonesIniciales };
